job-ats-matcher/requirements.txt
# Airflow - compatible version
apache-airflow==2.9.3

# Streamlit  
streamlit

# Data processing
pandas
numpy

# PDF processing
PyPDF2

# API requests
requests

# Environment variables
python-dotenv

# Utilities
python-dateutil

# Fix dependency conflicts
fastapi>=0.115.0